refactor log
