--- client/grpcclient/execute.go
@@ -1,1 +1,0 @@
-package grpcclient
